+++ src/user/user_model.h
@@ -216,7 +216,8 @@ class mjCModel {
   int nu;                         // number of actuators/controls
   int na;                         // number of activation variables
   int nmeshvert;                  // number of vertices in all meshes
-  int nmeshtexvert;               // number of texture coordinates in all meshes
+  int nmeshnormal;                // number of normals in all meshes
+  int nmeshtexcoord;              // number of texture coordinates in all meshes
   int nmeshface;                  // number of triangular faces in all meshes
   int nmeshgraph;                 // number of shorts in mesh auxiliary data
   int nskinvert;                  // number of vertices in all skins